requirements.txt
pystray~=0.19.5
pypresence~=4.3.0
pillow~=10.3.0
discord.py